+++ ChangeLog.md
@@ -7,6 +7,7 @@
     Release date: 
 
     - Define new TTP alertsWithFile(), alertsWithMms(), alertsWithMqtt(), alertsWithSms(), alertsWithSmtp(), alertsWithTts() functions
+    - ttp.pl alert takes advantage of these new functions
 
 ### 4.25.0